Increase line coverage of package prbt_hardware_support -> 100%
- move getParam into separate header file
- add getParam unit test

(cherry picked from commit 58157454caadc50950872dd349c86903b332a91e)
